recovery: add pacer
Currently we update the timestamp for next packet
(`last_packet_scheduled_time`) for each packet based on
the current pacing rate, which can confuse the
rtt measurement because the recorded timestamp
and actual time sent can be different, especially when
the application sends in burst using send_quantum
and batches a group packets using GSO.

Pacer provides the timestamp for next packet to be sent
based on the current send_quantum, pacing rate
and last updated time. It is reset when a valid packet is
received (including ACK and non-ACK packets) to
start a new burst.